fix: change data_length parameter type from uint16_t to size_t in build_cmd_frame for LD2410S

+++ components/ld2410s/LD2410S.cpp
@@ -64,7 +64,7 @@ namespace esphome
             this->send_command(read_config_cmd);
         }
 
-        CmdFrameT LD2410S::build_cmd_frame(uint16_t command, const uint8_t *data, uint16_t data_length)
+        CmdFrameT LD2410S::build_cmd_frame(uint16_t command, const uint8_t *data, size_t data_length)
         {
             CmdFrameT cmd_frame = {
                 .header = CMD_FRAME_HEADER,
